fix: f32 bug in edit ingredient route

+++ backend/src/routes/edit_ingredient.rs
@@ -52,7 +52,7 @@ pub struct IngredientPayload {
     pub category: String,
 
     #[serde(rename = "Quantity")]
-    pub quantity: u16,
+    pub quantity: f32,
 
     #[serde(rename = "Unit")]
     pub unit: String,
